% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rmarkdown.R
\name{add_css_for_rmd}
\alias{add_css_for_rmd}
\title{Add CSS file for RMarkdown to project directory}
\usage{
add_css_for_rmd(path = "reports/rmarkdown/www/rmd.css")
}
\arguments{
\item{path}{Default to 'reports/rmarkdown/www/rmd.css'.}
}
\description{
Add CSS file for RMarkdown to project directory
}
