% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{.onAttach}
\alias{.onAttach}
\title{On package attach, set ggplot theme and scales, for convenience}
\usage{
.onAttach(...)
}
\arguments{
\item{...}{(none)}
}
\description{
Set themes and scales for interactive sessions, projects, Shiny runtime, etc. For settings scales, see \link{theme_ggplot_raw} for details. NOTE cannot use \code{.onDetach <- function(...) { detach('env_theme_ggplot') }} as it prevents checking/installing for {theme} in the same session, though it works in a \code{--vanilla} session. TODO So should I use it?
}
\note{
Results in NAMESPACE \code{export()}. Doesn't need to be exported? link: https://community.rstudio.com/t/adding-ggplot-themes-and-color-palettes-to-a-package/2418/2
}
