% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_switcher.R
\name{get_theme_vals_ls}
\alias{get_theme_vals_ls}
\title{Get 'theme_vals' list}
\usage{
get_theme_vals_ls(theme)
}
\arguments{
\item{theme}{Default to 'elegant_dark' or, when shiny is running, 'theme_switcher_val' for modifying themes of ggplot/plotly/etc based on shiny theme. See \code{shinyr::theme_switcher} for details}
}
\description{
Return both name of theme and 'theme_vals' for \code{plotly_raw}, \code{theme_ggplot_raw}, etc
}
