% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{image_}
\alias{image_}
\title{Plot image scales}
\usage{
image_(obj, name = deparse(substitute(obj)))
}
\arguments{
\item{obj}{(none)}

\item{name}{(none)}
}
\description{
(none)
}
\examples{
\dontrun{
  rwb <- colorRampPalette(colors = c("red", "white", "blue"))
  img(pals_elegant_dark()(100), 'my_theme')
}

}
\references{
\href{https://stackoverflow.com/questions/37482977/what-is-a-good-palette-for-divergent-colors-in-r-or-can-viridis-and-magma-b/44553670}{What is a “good” palette for divergent colors in R? (or: can viridis and magma be combined together?)}
}
