% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vals.R
\name{pals}
\alias{pals}
\alias{pals_elegant_dark}
\alias{pals_elegant_light}
\alias{pals_elegant_light_2}
\title{Return function to interpolate a color palette}
\usage{
pals_elegant_dark(palette = "main_light", reverse = FALSE, n, ...)

pals_elegant_light(palette = "main", reverse = FALSE, n, ...)

pals_elegant_light_2(palette = "main", reverse = FALSE, n, ...)
}
\arguments{
\item{palette}{Character name of palette in palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{n}{(none)}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a color palette
}
\references{
\href{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}{Creating corporate colour palettes for ggplot2}
}
