% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly_raw}
\alias{plotly_raw}
\title{\code{plotly_raw}}
\usage{
plotly_raw(
  ...,
  theme_vals_ls,
  theme_vals = theme_vals_ls$theme_vals,
  theme_ = theme_vals_ls$theme,
  hovermode = "x",
  showspikes = TRUE,
  spikethickness = 1,
  font_col = theme_vals("col_7"),
  plot_bg = theme_vals("bg"),
  paper_bg = theme_vals("bg"),
  line_color = theme_vals("border"),
  axis_ticks = theme_vals("col_5"),
  grid_color = theme_vals("border"),
  colorway = theme_vals("col", "col_2", "col_4", "col_5", "fg", "col_9", "col_10",
    "col_11", "col_12"),
  colors = theme_vals("plotly_scales")
)
}
\arguments{
\item{...}{(none)}

\item{theme_vals_ls}{(none)}

\item{theme_vals}{(none)}

\item{theme_}{(none)}

\item{hovermode}{(none)}

\item{showspikes}{(none)}

\item{spikethickness}{(none)}

\item{font_col}{(none)}

\item{plot_bg}{(none)}

\item{paper_bg}{(none)}

\item{line_color}{(none)}

\item{axis_ticks}{(none)}

\item{grid_color}{(none)}

\item{colorway}{(none)}

\item{colors}{(none)}
}
\description{
\code{plotly_raw}
}
