% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly_shared_themes}
\alias{plotly_shared_themes}
\title{Shared plotly themes for \code{theme::ggplotly} & \code{theme::plotly}}
\usage{
plotly_shared_themes(
  p,
  theme_vals,
  rangeselector_bg = theme_vals("col_5_2"),
  rangeselector_bg_active = theme_vals("col_5"),
  rangeselector_fg = theme_vals("col_7")
)
}
\arguments{
\item{p}{(none)}

\item{theme_vals}{(none)}

\item{rangeselector_bg}{(none)}

\item{rangeselector_bg_active}{(none)}

\item{rangeselector_fg}{(none)}
}
\description{
Set default themes. To modify shared themes, use syntax of {plotly}
}
\note{
Call this as last step (don't know how to assign object back to calling frame)
}
