% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_func.R
\name{renderFunc}
\alias{renderFunc}
\alias{renderPlot}
\alias{renderPlotly}
\title{Custom shiny rendering functions}
\usage{
renderPlot(expr, ...)

renderPlotly(expr, ...)
}
\arguments{
\item{expr}{\emph{theme_switcher_val} from {shinyr} is called for updating plot after switching theme on input.}

\item{...}{(none)}
}
\description{
Wrapper for theming \code{shiny::renderPlot} & \code{plotly::renderPlotly}
}
\note{
Wrapper for \code{plotly::renderPlotly} doesn't work, hence \code{theme::renderPlotly} is used.

How does \code{theme::renderPlotly} updates ggplot theme from \code{theme::renderGGplot}

Use \code{conflict_prefer}
}
