% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vals.R
\name{scales}
\alias{scales}
\alias{scale_color_elegant_dark_d}
\alias{scale_color_elegant_dark_c}
\alias{scale_fill_elegant_dark_d}
\alias{scale_fill_elegant_dark_c}
\alias{scale_line_elegant_dark}
\alias{scale_color_elegant_light_d}
\alias{scale_color_elegant_light_c}
\alias{scale_fill_elegant_light_d}
\alias{scale_fill_elegant_light_c}
\alias{scale_line_elegant_light}
\alias{scale_color_elegant_light_2_d}
\alias{scale_color_elegant_light_2_c}
\alias{scale_fill_elegant_light_2_d}
\alias{scale_fill_elegant_light_2_c}
\alias{scale_line_elegant_light_2}
\title{Color and fill scale constructor for colors}
\usage{
scale_color_elegant_dark_d(palette = "main_light", reverse = FALSE, n, ...)

scale_color_elegant_dark_c(palette = "main_light", reverse = FALSE, n, ...)

scale_fill_elegant_dark_d(palette = "main_light", reverse = FALSE, n, ...)

scale_fill_elegant_dark_c(palette = "main_light", reverse = FALSE, n, ...)

scale_line_elegant_dark(n = 6)

scale_color_elegant_light_d(palette = "main", reverse = FALSE, n, ...)

scale_color_elegant_light_c(palette = "main", reverse = FALSE, n, ...)

scale_fill_elegant_light_d(palette = "main", reverse = FALSE, n, ...)

scale_fill_elegant_light_c(palette = "main", reverse = FALSE, n, ...)

scale_line_elegant_light(n = 6)

scale_color_elegant_light_2_d(palette = "main", reverse = FALSE, n, ...)

scale_color_elegant_light_2_c(palette = "main", reverse = FALSE, n, ...)

scale_fill_elegant_light_2_d(palette = "main", reverse = FALSE, n, ...)

scale_fill_elegant_light_2_c(palette = "main", reverse = FALSE, n, ...)

scale_line_elegant_light_2(n = 9)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{n}{(none)}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color and fill scale constructor for colors
}
