% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{suppress_plotly_msg}
\alias{suppress_plotly_msg}
\title{Suppress plotly messages}
\usage{
suppress_plotly_msg(p)
}
\arguments{
\item{p}{(none)}
}
\description{
Most plotly messages and warnings happen at print time, so if you want/need a more general way to suppress this (or other) messages (or warnings), you'll have to wrap a call to plotly_build(). Provide a parameter to suppress the "No trace type specified" warning+message](https://github.com/ropensci/plotly/issues/1202)
}
\note{
Performance of \code{plotly_build} is significant?
}
