% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import.R
\docType{package}
\name{theme-package}
\alias{theme}
\alias{theme-package}
\title{theme: Themes for ggplot2, Shiny, Rmarkdown, etc}
\description{
Set ggplot theme and scales when package is attached, for convenience. Personalized themes, scales, theme switcher for Shiny, Rmarkdown, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kar9222/theme}
  \item Report bugs at \url{https://github.com/kar9222/theme/issues}
}

}
\author{
\strong{Maintainer}: Kar \email{first.last@example.com}

}
