% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_ggplot}
\alias{theme_ggplot}
\alias{theme_ggplot_elegant_dark}
\alias{theme_ggplot_elegant_light}
\alias{theme_ggplot_elegant_light_2}
\alias{theme_ggplot_raw}
\title{Custom ggplot theme for geom & layout}
\usage{
theme_ggplot(..., theme)

theme_ggplot_elegant_dark(...)

theme_ggplot_elegant_light(...)

theme_ggplot_elegant_light_2(...)

theme_ggplot_raw(
  theme_vals_ls,
  theme_vals = theme_vals_ls$theme_vals,
  theme_ = theme_vals_ls$theme,
  strip_text_face = "bold",
  geom_bar_fill = theme_vals("col_5_3"),
  geom_col_col = theme_vals("col_5_3"),
  geom_col_fill = theme_vals("col_5_3"),
  geom_point_col = theme_vals("col"),
  geom_line_col = theme_vals("col"),
  geom_smooth_col = theme_vals("col"),
  geom_smooth_fill = theme_vals("col_5"),
  geom_area_col = theme_vals("col_5_3"),
  geom_area_fill = theme_vals("col_5_3"),
  geom_density_col = theme_vals("col"),
  geom_ribbon_fill = theme_vals("col_5"),
  geom_ribbon_alpha = 0.3,
  geom_hline_col = theme_vals("col_5"),
  geom_hline_linetype = "dashed",
  geom_vline_col = theme_vals("col_5"),
  geom_vline_linetype = "dashed",
  geom_segment_col = theme_vals("col_2"),
  geom_text_col = theme_vals("col_13"),
  geom_slabinterval_fill = theme_vals("col_5_3"),
  geom_slabinterval_interval_colour = theme_vals("col_13"),
  geom_slabinterval_point_colour = theme_vals("col_13"),
  geom_lineribbon_colour = theme_("selection_2"),
  geom_dagedgepath_edge_colour = theme_("col_5_3"),
  geom_dagpoint_colour = theme_("col_5_2"),
  plot_background = theme_vals("bg"),
  plot_background_border = theme_vals("bg"),
  panel_background = theme_vals("bg"),
  panel_background_border = theme_vals("border"),
  panel_grid = theme_vals("border"),
  title = theme_vals("col_7"),
  plot_subtitle = theme_vals("col_7"),
  plot_caption = theme_vals("col_5"),
  plot_tag = theme_vals("col_5"),
  axis_line = theme_vals("border"),
  axis_title = theme_vals("col_7"),
  axis_text = theme_vals("col_7"),
  axis_ticks = theme_vals("col_5"),
  legend_title = theme_vals("col_7"),
  legend_text = theme_vals("col_7"),
  strip_background = theme_vals("selection"),
  strip_background_border = theme_vals("selection"),
  strip_text = theme_vals("col_7")
)
}
\arguments{
\item{...}{(none)}

\item{theme}{For internal use}

\item{theme_vals_ls}{(none)}

\item{theme_vals}{(none)}

\item{theme_}{(none)}

\item{strip_text_face}{(none)}

\item{geom_bar_fill}{(none)}

\item{geom_col_col}{(none)}

\item{geom_col_fill}{(none)}

\item{geom_point_col}{(none)}

\item{geom_line_col}{(none)}

\item{geom_smooth_col}{(none)}

\item{geom_smooth_fill}{(none)}

\item{geom_area_col}{(none)}

\item{geom_area_fill}{(none)}

\item{geom_density_col}{(none)}

\item{geom_ribbon_fill}{(none)}

\item{geom_ribbon_alpha}{(none)}

\item{geom_hline_col}{(none)}

\item{geom_hline_linetype}{(none)}

\item{geom_vline_col}{(none)}

\item{geom_vline_linetype}{(none)}

\item{geom_segment_col}{(none)}

\item{geom_text_col}{(none)}

\item{geom_slabinterval_fill}{(none)}

\item{geom_slabinterval_interval_colour}{(none)}

\item{geom_slabinterval_point_colour}{(none)}

\item{geom_lineribbon_colour}{(none)}

\item{geom_dagedgepath_edge_colour}{(none)}

\item{geom_dagpoint_colour}{(none)}

\item{plot_background}{(none)}

\item{plot_background_border}{(none)}

\item{panel_background}{(none)}

\item{panel_background_border}{(none)}

\item{panel_grid}{(none)}

\item{title}{(none)}

\item{plot_subtitle}{(none)}

\item{plot_caption}{(none)}

\item{plot_tag}{(none)}

\item{axis_line}{(none)}

\item{axis_title}{(none)}

\item{axis_text}{(none)}

\item{axis_ticks}{(none)}

\item{legend_title}{(none)}

\item{legend_text}{(none)}

\item{strip_background}{(none)}

\item{strip_background_border}{(none)}

\item{strip_text}{(none)}
}
\description{
Available themes are 'elegant_dark', 'elegant_light' and 'elegant_light_2'. \code{theme_switcher_val} For theme ggplot in shiny. See \code{shinyr::theme_switcher} for details. This argument is used for resolving specific behaviour (ie env of shiny?) in Shiny.
}
\details{
Attach custom environment named 'env_theme_ggplot' to search path for
\itemize{
\item for getting name of theme in \code{theme::geom_line}, etc by calling \code{theme_ggplot_name}
\item due to complexities of environments and Shiny runtime, new environment named 'env_theme_ggplot' is set for overriding and storing ggplot scales objects upon \code{.onAttach} and \code{.onDetach}.
}

Also see \link{viridis_scales_from_ggplot2}
}
