% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{theme_shiny}
\alias{theme_shiny}
\title{Shiny theme}
\usage{
theme_shiny(
  theme_vals,
  button_height = 34,
  button_padding = "6px 12px",
  textbox_height = 34,
  textbox_padding = "6px 12px",
  logo_text_font_family = "Consolas",
  logo_text_font_size = "25px !important",
  logo_text_font_style = "italic",
  transition = ".2s",
  app_font_family = "Arial",
  app_font = theme_vals("fg_3"),
  primary_font = "rgb(0,0,0)",
  info_font = "rgb(0,0,0)",
  soccess_font = "rgb(0,0,0)",
  warning_font = "rgb(0,0,0)",
  danger_font = "rgb(0,0,0)",
  body_bg = theme_vals("bg_2"),
  placeholder_fg = theme_vals("col_5"),
  selection = theme_vals("selection"),
  search_box_input_fg = theme_vals("col_2"),
  header_bg = theme_vals("bg_3"),
  logo_bg = theme_vals("bg_3"),
  logo_fg = theme_vals("col_5_3"),
  logo_hover_fg = theme_vals("hover_fg"),
  header_button_bg = theme_vals("bg_3"),
  header_button_icon_col = theme_vals("col_5_3"),
  header_button_hover_bg = theme_vals("bg_3"),
  header_button_icon_hover_fg = theme_vals("hover_fg"),
  header_dropdown_fg = theme_vals("col_5_3"),
  header_dropdown_hover_bg = theme_vals("bg_3"),
  header_dropdown_hover_fg = theme_vals("hover_fg"),
  header_dropdown_focus_bg = theme_vals("bg_3"),
  header_dropdown_focus_fg = theme_vals("hover_fg"),
  header_box_shadow = theme_vals("bg"),
  header_box_shadow_size = "2px 2px 2px",
  sidebar_bg = theme_vals("bg_3"),
  sidebar_padding = 0,
  sidebar_menu_bg = "transparent",
  sidebar_menu_padding = 0,
  sidebar_menu_border_radius = 0,
  sidebar_shadow_radius = "3px 5px 5px",
  sidebar_shadow = theme_vals("bg_3"),
  sidebar_user_fg = theme_vals("fg_2"),
  sidebar_search_bg = theme_vals("bg_3"),
  sidebar_search_icon_col = theme_vals("col_5"),
  sidebar_search_border_col = theme_vals("bg_3"),
  sidebar_tab_fg = theme_vals("fg_3"),
  sidebar_tab_text_size = 13,
  sidebar_tab_border_style = "none none solid none",
  sidebar_tab_border_col = theme_vals("shiny_sidebar_border"),
  sidebar_tab_border_width = 1,
  sidebar_tab_selected_bg = theme_vals("bg_3"),
  sidebar_tab_selected_fg = theme_vals("selected_fg"),
  sidebar_tab_selected_radius = "0px 0px 0px 0px",
  sidebar_tab_hover_bg = theme_vals("hover_bg"),
  sidebar_tab_hover_fg = theme_vals("hover_fg"),
  sidebar_tab_border_hover_style = "none none solid none",
  sidebar_tab_border_hover_col = theme_vals("border"),
  sidebar_tab_border_hover_width = 1,
  sidebar_tab_hover_radius = "0px 0px 0px 0px",
  box_outer_bg = theme_vals("bg_box"),
  box_bg = theme_vals("bg_box"),
  box_border_radius = 1,
  box_shadow_size = theme_vals("shadow_size"),
  box_shadow_col = theme_vals("shadow_bg"),
  box_title_size = 18,
  box_default_col = theme_vals("bg_box"),
  box_primary_col = "rgb(15,124,191)",
  box_info_col = "rgb(225,225,225)",
  box_success_col = "rgb(59,133,95)",
  box_warning_col = "rgb(178,83,149)",
  box_danger_col = "rgb(207,57,92)",
  tab_box_tab_col = theme_vals("bg_box"),
  tab_box_tab_text_size = 14,
  tab_box_tab_text_col = "rgb(42,102,98)",
  tab_box_tab_text_selected_fg = "rgb(207,57,92)",
  tab_box_bg = theme_vals("bg"),
  tab_box_highlight = "rgb(207,57,92)",
  tab_box_border_radius = 0,
  input_arrow_display = "none",
  button_bg = theme_vals("bg_box"),
  button_fg = theme_vals("shiny_box_content_fg"),
  button_border_col = theme_vals("bg_box"),
  button_border_radius = 0,
  button_hover_bg = theme_vals("hover_bg_2"),
  button_hover_fg = theme_vals("fg_3"),
  button_border_hover_col = theme_vals("bg_box"),
  material_switch_bg = theme_vals("col_5"),
  material_switch_before_bg = theme_vals("col_13_3"),
  material_switch_after_bg = theme_vals("col_4_2"),
  textbox_bg = theme_vals("bg_box"),
  textbox_border_col = theme_vals("bg_box"),
  textbox_border_radius = 0,
  textbox_selected_bg = theme_vals("selected_bg"),
  textbox_border_selected_col = theme_vals("bg_box"),
  table_bg = theme_vals("bg"),
  table_fg = theme_vals("border"),
  table_border_top_size = 1,
  table_border_row_size = 1,
  selectize_input_bg = theme_vals("bg"),
  selectize_input_fg = theme_vals("shiny_box_content_fg"),
  selectize_dropdown_hover_bg = theme_vals("hover_bg_2"),
  selectize_dropdown_hover_fg = theme_vals("fg_3"),
  selectize_dropdown_bg = theme_vals("bg"),
  selectize_dropdown_fg = theme_vals("fg_3"),
  selectize_dropdown_active_bg = theme_vals("hover_bg"),
  selectize_dropdown_active_fg = theme_vals("fg_3"),
  selectize_dropdown_highlight_bg = theme_vals("bg"),
  selectize_dropdown_highlight_fg = theme_vals("col_2"),
  selectize_input_border = theme_vals("border"),
  selectize_dropdown_border = theme_vals("border"),
  selectize_input_others_bg = theme_vals("bg"),
  date_range_input_fg = theme_vals("shiny_box_content_fg"),
  date_range_dropdown_border_col = theme_vals("border"),
  date_range_dropdown_bg = theme_vals("bg"),
  date_range_input_active_bg = theme_vals("bg"),
  date_range_input_active_fg = theme_vals("shiny_box_content_fg"),
  date_range_input_hover_bg = theme_vals("bg"),
  date_range_input_hover_fg = theme_vals("shiny_box_content_fg"),
  date_range_others_active_bg = theme_vals("bg"),
  date_range_others_active_fg = theme_vals("shiny_box_content_fg"),
  date_range_others_active_border_col = "none",
  date_range_others_hover_bg = theme_vals("bg"),
  date_range_others_hover_fg = theme_vals("shiny_box_content_fg"),
  date_range_to_middle_part_bg = theme_vals("bg"),
  date_range_to_middle_part_others_fg = theme_vals("shiny_box_content_fg"),
  date_range_to_middle_part_others_bg = theme_vals("bg"),
  date_range_input_middle_button_fg = theme_vals("col_5"),
  date_range_year_choices_fg = theme_vals("fg_3"),
  date_range_year_choices_hover_bg = theme_vals("bg"),
  reactable_bg = theme_vals("bg"),
  reactable_bg_strip = theme_vals("col_8"),
  bg = theme_vals("bg"),
  fg = theme_vals("fg_3"),
  fg_2 = theme_vals("fg_2"),
  shadow_bg = theme_vals("shadow_bg"),
  shadow_size = theme_vals("shadow_size"),
  hover_bg = theme_vals("hover_bg"),
  hover_fg = theme_vals("hover_fg"),
  hover_bg_table = theme_vals("hover_bg_table"),
  selected_bg = theme_vals("selected_bg"),
  selected_fg = theme_vals("selected_fg"),
  bg_box = theme_vals("bg_box"),
  border = theme_vals("border"),
  border_2 = theme_vals("border_2")
)
}
\arguments{
\item{theme_vals}{(none)}

\item{button_height}{(none)}

\item{button_padding}{(none)}

\item{textbox_height}{(none)}

\item{textbox_padding}{(none)}

\item{logo_text_font_family}{(none)}

\item{logo_text_font_size}{(none)}

\item{logo_text_font_style}{(none)}

\item{transition}{(none)}

\item{app_font_family}{(none)}

\item{app_font}{(none)}

\item{primary_font}{(none)}

\item{info_font}{(none)}

\item{soccess_font}{(none)}

\item{warning_font}{(none)}

\item{danger_font}{(none)}

\item{body_bg}{(none)}

\item{placeholder_fg}{(none)}

\item{selection}{(none)}

\item{search_box_input_fg}{(none)}

\item{header_bg}{(none)}

\item{logo_bg}{(none)}

\item{logo_fg}{(none)}

\item{logo_hover_fg}{(none)}

\item{header_button_bg}{(none)}

\item{header_button_icon_col}{(none)}

\item{header_button_hover_bg}{(none)}

\item{header_button_icon_hover_fg}{(none)}

\item{header_dropdown_fg}{(none)}

\item{header_dropdown_hover_bg}{(none)}

\item{header_dropdown_hover_fg}{(none)}

\item{header_dropdown_focus_bg}{(none)}

\item{header_dropdown_focus_fg}{(none)}

\item{header_box_shadow}{(none)}

\item{header_box_shadow_size}{(none)}

\item{sidebar_bg}{(none)}

\item{sidebar_padding}{(none)}

\item{sidebar_menu_bg}{(none)}

\item{sidebar_menu_padding}{(none)}

\item{sidebar_menu_border_radius}{(none)}

\item{sidebar_shadow_radius}{(none)}

\item{sidebar_shadow}{(none)}

\item{sidebar_user_fg}{(none)}

\item{sidebar_search_bg}{(none)}

\item{sidebar_search_icon_col}{(none)}

\item{sidebar_search_border_col}{(none)}

\item{sidebar_tab_fg}{(none)}

\item{sidebar_tab_text_size}{(none)}

\item{sidebar_tab_border_style}{(none)}

\item{sidebar_tab_border_col}{(none)}

\item{sidebar_tab_border_width}{(none)}

\item{sidebar_tab_selected_bg}{(none)}

\item{sidebar_tab_selected_fg}{(none)}

\item{sidebar_tab_selected_radius}{(none)}

\item{sidebar_tab_hover_bg}{(none)}

\item{sidebar_tab_hover_fg}{(none)}

\item{sidebar_tab_border_hover_style}{(none)}

\item{sidebar_tab_border_hover_col}{(none)}

\item{sidebar_tab_border_hover_width}{(none)}

\item{sidebar_tab_hover_radius}{(none)}

\item{box_outer_bg}{(none)}

\item{box_bg}{(none)}

\item{box_border_radius}{(none)}

\item{box_shadow_size}{(none)}

\item{box_shadow_col}{(none)}

\item{box_title_size}{(none)}

\item{box_default_col}{(none)}

\item{box_primary_col}{(none)}

\item{box_info_col}{(none)}

\item{box_success_col}{(none)}

\item{box_warning_col}{(none)}

\item{box_danger_col}{(none)}

\item{tab_box_tab_col}{(none)}

\item{tab_box_tab_text_size}{(none)}

\item{tab_box_tab_text_col}{(none)}

\item{tab_box_tab_text_selected_fg}{(none)}

\item{tab_box_bg}{(none)}

\item{tab_box_highlight}{(none)}

\item{tab_box_border_radius}{(none)}

\item{input_arrow_display}{(none)}

\item{button_bg}{(none)}

\item{button_fg}{(none)}

\item{button_border_col}{(none)}

\item{button_border_radius}{(none)}

\item{button_hover_bg}{(none)}

\item{button_hover_fg}{(none)}

\item{button_border_hover_col}{(none)}

\item{material_switch_bg}{(none)}

\item{material_switch_before_bg}{(none)}

\item{material_switch_after_bg}{(none)}

\item{textbox_bg}{(none)}

\item{textbox_border_col}{(none)}

\item{textbox_border_radius}{(none)}

\item{textbox_selected_bg}{(none)}

\item{textbox_border_selected_col}{(none)}

\item{table_bg}{(none)}

\item{table_fg}{(none)}

\item{table_border_top_size}{(none)}

\item{table_border_row_size}{(none)}

\item{selectize_input_bg}{(none)}

\item{selectize_input_fg}{(none)}

\item{selectize_dropdown_hover_bg}{(none)}

\item{selectize_dropdown_hover_fg}{(none)}

\item{selectize_dropdown_bg}{(none)}

\item{selectize_dropdown_fg}{(none)}

\item{selectize_dropdown_active_bg}{(none)}

\item{selectize_dropdown_active_fg}{(none)}

\item{selectize_dropdown_highlight_bg}{(none)}

\item{selectize_dropdown_highlight_fg}{(none)}

\item{selectize_input_border}{(none)}

\item{selectize_dropdown_border}{(none)}

\item{selectize_input_others_bg}{(none)}

\item{date_range_input_fg}{(none)}

\item{date_range_dropdown_border_col}{(none)}

\item{date_range_dropdown_bg}{(none)}

\item{date_range_input_active_bg}{(none)}

\item{date_range_input_active_fg}{(none)}

\item{date_range_input_hover_bg}{(none)}

\item{date_range_input_hover_fg}{(none)}

\item{date_range_others_active_bg}{(none)}

\item{date_range_others_active_fg}{(none)}

\item{date_range_others_active_border_col}{(none)}

\item{date_range_others_hover_bg}{(none)}

\item{date_range_others_hover_fg}{(none)}

\item{date_range_to_middle_part_bg}{(none)}

\item{date_range_to_middle_part_others_fg}{(none)}

\item{date_range_to_middle_part_others_bg}{(none)}

\item{date_range_input_middle_button_fg}{(none)}

\item{date_range_year_choices_fg}{(none)}

\item{date_range_year_choices_hover_bg}{(none)}

\item{reactable_bg}{(none)}

\item{reactable_bg_strip}{(none)}

\item{bg}{(none)}

\item{fg}{(none)}

\item{fg_2}{(none)}

\item{shadow_bg}{(none)}

\item{shadow_size}{(none)}

\item{hover_bg}{(none)}

\item{hover_fg}{(none)}

\item{hover_bg_table}{(none)}

\item{selected_bg}{(none)}

\item{selected_fg}{(none)}

\item{bg_box}{(none)}

\item{border}{(none)}

\item{border_2}{(none)}
}
\description{
Shiny theme
}
