% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_switcher.R
\name{theme_switcher}
\alias{theme_switcher}
\alias{theme_switcher_shiny}
\title{Theme switcher}
\usage{
theme_switcher_shiny(theme, ...)
}
\arguments{
\item{theme}{(none)}

\item{...}{(none)}
}
\description{
Theme switcher for plotly and shiny.
}
\note{
\code{theme_func} argument of internal function \code{theme::theme_switcher}: choose one of 'shiny', 'plotly'
}
