% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vals.R
\name{theme_vals}
\alias{theme_vals}
\alias{theme_}
\alias{elegant_dark}
\alias{theme_vals_elegant_dark}
\alias{theme_vals_elegant_light}
\alias{theme_vals_elegant_light_2}
\title{Theme values for font colors, background colors, font styles, etc}
\usage{
theme_(...)

elegant_dark(...)

theme_vals_elegant_dark(...)

theme_vals_elegant_light(...)

theme_vals_elegant_light_2(...)
}
\arguments{
\item{...}{Names of theme_vals

#' @references For 'elegant_light', see \href{https://github.com/mtyn/polar}{polar}}
}
\description{
(none)
}
