% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{viridis_scales_from_ggplot2}
\alias{viridis_scales_from_ggplot2}
\alias{scale_color_viridis_c}
\alias{scale_color_viridis_d}
\alias{scale_fill_viridis_c}
\alias{scale_fill_viridis_d}
\title{Viridis scales e.g. scale_color_viridis_c from ggplot2}
\usage{
scale_color_viridis_c(...)

scale_color_viridis_d(...)

scale_fill_viridis_c(...)

scale_fill_viridis_d(...)
}
\arguments{
\item{...}{(none)}
}
\description{
Export these functions for \code{theme::theme_ggplot_2} so namespaces can be found by \code{match.fun}
}
