% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-package.r
\docType{package}
\name{theme-package}
\alias{theme}
\alias{theme-package}
\title{theme: Base R Plot Themes}
\description{
Provides a framework for creating and using base R plot themes.
}
\details{
Global plot settings
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

}
